Add abstract class reporting
